+++ CHANGELOG.md
@@ -38,6 +38,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.18] - 2026-03-29
+
+### Docs
+- Update examples/README.md
+- Update examples/llx/README.md
+
+### Other
+- Update examples/llx/demo.sh
+- Update examples/llx/pyqual-llx.yaml
+- Update pyqual/plugins.py
+
 ## [0.1.17] - 2026-03-29
 
 ### Docs